Created App class to hold main of GUI
refactored Animal imagePath to camel caseing. Linked list demo now shoes images.
